+++ spec/overcommit/configuration_validator_spec.rb
@@ -1,7 +1,9 @@
 require 'spec_helper'
 
 describe Overcommit::ConfigurationValidator do
-  let(:options) { {} }
+  let(:output) { StringIO.new }
+  let(:logger) { Overcommit::Logger.new(output) }
+  let(:options) { { logger: logger } }
   let(:config) { Overcommit::Configuration.new(config_hash, validate: false) }
 
   subject { described_class.new.validate(config, config_hash, options) }
@@ -19,6 +21,56 @@ describe Overcommit::ConfigurationValidator do
 
     it 'raises an error' do
       expect { subject }.to raise_error Overcommit::Exceptions::ConfigurationError
+    end
+  end
+
+  context 'when hook has `env` set' do
+    let(:config_hash) do
+      {
+        'PreCommit' => {
+          'MyHook' => {
+            'enabled' => true,
+            'env' => env,
+          },
+        },
+      }
+    end
+
+    context 'and it is a single string' do
+      let(:env) { 'OVERCOMMIT_ENV_VAR=1' }
+
+      it 'raises an error and mentions `env` must be a hash' do
+        expect { subject }.to raise_error Overcommit::Exceptions::ConfigurationError
+        output.string.should =~ /must be a hash/i
+      end
+    end
+
+    context 'and it is a hash with string values' do
+      let(:env) { { 'OVERCOMMIT_ENV_VAR' => '1', 'OVERCOMMIT_ENV_VAR_2' => '2' } }
+
+      it 'is valid' do
+        expect { subject }.not_to raise_error
+      end
+    end
+
+    context 'and it is a hash with integer values' do
+      let(:env) { { 'OVERCOMMIT_ENV_VAR' => 1, 'OVERCOMMIT_ENV_VAR_2' => 2 } }
+
+      it 'raises an error' do
+        expect { subject }.to raise_error Overcommit::Exceptions::ConfigurationError
+        output.string.should =~ /`OVERCOMMIT_ENV_VAR`.*must be a string/i
+        output.string.should =~ /`OVERCOMMIT_ENV_VAR_2`.*must be a string/i
+      end
+    end
+
+    context 'and it is a hash with boolean values' do
+      let(:env) { { 'OVERCOMMIT_ENV_VAR' => true, 'OVERCOMMIT_ENV_VAR_2' => false } }
+
+      it 'raises an error' do
+        expect { subject }.to raise_error Overcommit::Exceptions::ConfigurationError
+        output.string.should =~ /`OVERCOMMIT_ENV_VAR`.*must be a string/i
+        output.string.should =~ /`OVERCOMMIT_ENV_VAR_2`.*must be a string/i
+      end
     end
   end